moving tests to existing file
